simulation/requirements.txt
taichi
plyfile
numpy 
PyMCubes